open_meteo_API/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
geographiclib==2.0
geopy==2.3.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pydantic==1.10.7
pytest==7.3.1
pytest-django==4.5.2
pytest-mock==3.10.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
